+++ examples/cpi/simple-cpi-6/trident-tests/fuzz_0/instructions/initialize_callee.rs
@@ -10,6 +10,7 @@ pub struct InitializeCalleeInstruction {
 }
 /// Instruction Accounts
 #[derive(Arbitrary, Debug, Clone, TridentAccounts)]
+#[instruction_data(InitializeCalleeInstructionData)]
 pub struct InitializeCalleeInstructionAccounts {
     pub signer: TridentAccount,
 }