+++ examples/cpi/simple-cpi-6/trident-tests/fuzz_0/instructions/initialize_caller.rs
@@ -10,6 +10,7 @@ pub struct InitializeCallerInstruction {
 }
 /// Instruction Accounts
 #[derive(Arbitrary, Debug, Clone, TridentAccounts)]
+#[instruction_data(InitializeCallerInstructionData)]
 pub struct InitializeCallerInstructionAccounts {
     #[account(signer, mut,storage = signer)]
     pub signer: TridentAccount,